requirements.txt
ollama
litellm
agentops
openai
openai-agents  # OpenAI Agents SDK
python-dotenv
requests
pydantic